Added more dashboard apis
Api added for pie chart which will get active vs complete tasks, based on a project id, which will need a dropdown for all projects. May have this in the projects overview for each project. Not sure yet. 

Also need to add the extra fields for Task table into the Task model class for the completionDate